requirements.txt
absl-py==0.13.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.0.5
async-generator==1.10
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.9.3
bleach==5.0.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
colorama==0.4.4
cryptography==37.0.4
cv==1.0.0
cycler==0.10.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.3
Django==3.2.9
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
djangorestframework==3.13.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.4.0
fonttools==4.33.3
h11==0.13.0
idna==3.2
ipykernel==6.13.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.6.0
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keyboard==0.13.5
kiwisolver==1.3.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mediapipe==0.8.6
mistune==0.8.4
MouseInfo==0.1.3
mysql-connector==2.2.9
mysql-connector-python==8.0.27
mysqlclient==2.1.0
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.4
opencv-contrib-python==4.5.3.56
outcome==1.2.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.3.1
platformdirs==2.4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.17.3
psutil==5.9.1
pure-eval==0.2.2
PyAutoGUI==0.9.53
pycodestyle==2.8.0
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.12.0
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyperclip==1.8.2
pypyodbc==1.3.2
PyRect==0.1.4
pyrsistent==0.18.1
PyScreeze==0.1.27
PySocks==1.7.1
python-dateutil==2.8.2
PyTweening==1.0.3
pytz==2021.3
pywin32==304
pywinpty==2.0.5
pyzmq==23.1.0
qtconsole==5.3.0
QtPy==2.1.0
requests==2.26.0
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
selenium==4.3.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.2
stack-data==0.2.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tornado==6.1
traitlets==5.2.2.post1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.6
virtualenv==20.10.0
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
wsproto==1.1.0
xlrd==1.2.0
